project/requirements.txt
lmstudio
pandas
google-genai
python-dotenv
nltk
textstat
sentence_transformers
jupyter